rag_qa_assistant/requirements.txt
langchain==0.1.14
langchain-community>=0.0.9
langchain-openai>=0.0.8
openai==1.30.1
faiss-cpu== 1.11.0
tiktoken==0.7.0
python-dotenv==1.0.1
